requirements.txt
Flask
Flask-SQLAlchemy
groq
python-dotenv
email-validator
requests
pytz
